+++ src/experiments/policy_transfer.py
@@ -37,7 +37,7 @@ from functools import partial
 import configparser
 
 from src.train import get_parser, main
-from src.context.context_sampler import get_default_context_and_bounds
+from src.context.sampling import get_default_context_and_bounds
 from src.envs import CARLVehicleRacingEnv, CARLLunarLanderEnv
 from src.envs.box2d.meta_vehicle_racing import RaceCar, AWDRaceCar, StreetCar, TukTuk, BusSmallTrailer, PARKING_GARAGE
